Adds A Fallback Accessibility Label to Number Forms
This relates to one of the items (Age text field) in #3151

By giving the number label a default value when it would otherwise have
none, we provide helpful context when tapping into the field. This
seemed applicable to the accessibility of all fields, so I’ve placed
the pull request here so that every number form cell can receive this
accessibility benefit.